Split account storage by what it is about
Two thousand lines in one file, of which the schema and its migrations
were five hundred sitting between the queries that use them, so reading
create_session meant scrolling past every CREATE TABLE in the product.
The module root keeps the types, the shared connection and the login
helpers every query needs and re-exports the rest, so a caller still
writes accounts::create_session and does not learn which file it landed
in. Behaviour is unchanged: the only edits to moved code are the
visibility keywords the compiler asked for.
